feat: introduce config registry manager and publisher with enhanced snapshot handling and validation
